Split target runtimes from UnitTestAndCompileExtension
Target runtimes, like usedInWorkers and usedInDaemon, are used to calculate the JVM version that a project
should compile to. In the future, these values will be used to enforce target runtime boundaries: e.g.
something usedInWorkers must only depend on other things that are usedInWorkers.

However, some projects are not part of the distribution, like testing infrastructure, and therefore it
does not make sense to mark them as used in some target runtime -- they are only used during the build.

We already have plugins for these types of projects -- gradlebuild.internal.java and gradlebuild.internal.kotlin.
These plugins mark projects that are 'internal' as in they do not contribute to the distribution.

This commit makes a number of changes
- Pulls usedInWorkers, usedInClient -- AKA the target runtime definitions -- into the 'targetRuntimes' block of a new extension, the GradleModuleExtension
- Updates the UnitTestAndCompileExtension to accept a 'targetJvmVersion' property
- Updates 'targetJvmVersion', as a convention, to be configured based on the values of 'targetRuntimes', as was before, allowing this value to be specified in ways different than setting targetRuntimes
- Updates gradlebuild.internal.java and gradlebuild.internal.kotlin to hard-code their target versions, instad of needing to declare their target runtimes.
- Merges the existing ModuleIdentityExtension to an 'identity' block of the new GradleModuleExtension -- ensuring the new extension is roughly compatible with a 'declarative-like' style that will eventually be declarative-gradle compatible
- Updates existing logic that computes module identity versions in the 'identity' block to be provider-aware instead of being computed eagerly.